+++ app/Composers/CurrencyComposer.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Composers;
+
+use App\Models\Currency;
+use Illuminate\View\View;
+
+class CurrencyComposer
+{
+    public function compose(View $view)
+    {
+        $view->with('currencies',Currency::all());
+    }
+}