+++ database/seeders/DatabaseSeeder.php
@@ -2,6 +2,8 @@
 
 namespace Database\Seeders;
 
+use App\Models\Currency;
+use App\Models\Languages;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 
@@ -15,8 +17,11 @@ class DatabaseSeeder extends Seeder
     public function run()
     {
         $this->call([
+            LanguagesSeeder::class,
+            CurrencySeeder::class,
             CategorySeeder::class,
             ProductSeeder::class,
+            AdminSeeder::class,
         ]);
     }
 }